+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- New folder and Delete folder in both panes; delete confirms and sends the folder to MEGA/PikPak trash (#11)
+
 ### Changed
 
 ### Fixed